add Restart method + improve tests (#20)
